third_party/WebKit/LayoutTests/inspector/elements/styles/styles-add-new-rule-to-stylesheet.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script src="styles-test.js"></script>
<link rel="stylesheet" href="resources/different-rule-types.css">
<script>

function test()
{
    InspectorTest.waitForStylesheetsOnFrontend(2, onStyleSheetsLoaded);

    function onStyleSheetsLoaded()
    {
        var styleSheets = WebInspector.cssModel.allStyleSheets();
        for (var i = 0; i < styleSheets.length; ++i) {
            var styleSheetHeader = styleSheets[i];
            if (styleSheetHeader.sourceURL.indexOf("different-rule-types.css") === -1)
                continue;
            InspectorTest.addNewRuleInStyleSheet(styleSheetHeader, "#other, div", onRuleAdded);
        }
    }

    function onRuleAdded()
    {
        InspectorTest.selectNodeAndWaitForStyles("other", onNodeSelected);
    }

    function onNodeSelected()
    {
        InspectorTest.dumpSelectedElementStyles(true, false, true);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that adding new rule in the stylesheet end works as expected.
</p>

<div class="inspected">Styled element</div>
<div id="other"></div>

</body>
</html>
